+++ docs/source/CHANGELOG.md
@@ -48,4 +48,5 @@
 
 ### 1.4.0 (in progress)
 
+- Improve generality of the app test manager configurations and fixup docstrings
 - Set `TESTING` attribute to `False` as default for configurations