+++ docs/source/CHANGELOG.md
@@ -46,8 +46,9 @@
 
 - Remove double query for acquiring authorized entries
 
-### 1.4.0 (in progress)
+### 1.4.0
 
 - Generalize custom configuration parameters
+- Add a configuration parameter for a preloadable data callable
 - Improve generality of the app test manager configurations and fixup docstrings
 - Set `TESTING` attribute to `False` as default for configurations